third_party/blink/web_tests/wpt_internal/digital-credentials/identity-get.tentative.https.html

<!DOCTYPE html>
<title>Digital Identity Credential tests.</title>
<link rel="help" href="https://wicg.github.io/digital-identities/">
<script src="/common/get-host-info.sub.js"></script>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/resources/testdriver.js"></script>
<script src="/resources/testdriver-vendor.js"></script>

<body>
<script type="module">
  import { buildValidNavigatorCredentialsRequest, requestIdentityWithActivation } from '/wpt_internal/digital-credentials/support/helper.js';

// Internal test because the spec currently does not require transient user activation.
promise_test(async t=>{
  assert_false(navigator.userActivation.isActive);
  let request = buildValidNavigatorCredentialsRequest();
  await promise_rejects_dom(t, "NotAllowedError", navigator.credentials.get(request));
}, "navigator.credentials.get() fails if the page doesn't have user activation");

// Tests for success are here because the tests rely on the
// --use-fake-ui-for-digital-identity command line flag. Moving the tests to
// /external requires a WebDriver testing API.

promise_test(async t => {
  let request = buildValidNavigatorCredentialsRequest();
  let credential = await requestIdentityWithActivation(test_driver, request);
  let expectedResponseData = { "token": "fake_test_token" };
  assert_equals("openid4vp", credential.protocol);
  assert_equals(
    JSON.stringify(expectedResponseData),
    JSON.stringify(credential.data));
}, "navigator.credentials.get() API works in toplevel frame.");

promise_test(async t=> {
  let abortController = new AbortController();
  let request = buildValidNavigatorCredentialsRequest();
  request.signal = abortController.signal;
  let requestPromise = requestIdentityWithActivation(test_driver, request);
  abortController.abort();
  await promise_rejects_dom(t, "AbortError", requestPromise);
}, "navigator.credentials.get() promise is rejected when the page aborts the request.");
</script>
